### 2017-06-12 #292 @whedon
[PRE REVIEW]: cuIBM: A GPU-based immersed boundary method code
**Submitting author:** @mesnardo (<a href="http://orcid.org/0000-0001-5335-7853">Olivier Mesnard</a>)
**Repository:** <a href="https://github.com/barbagroup/cuIBM" target ="_blank">https://github.com/barbagroup/cuIBM</a>
**Version:** v0.1
