### 2017-06-18 #292 @kyleniemeyer
Hi @mesnardo, thanks for submitting cuIBM to JOSS!

(For full disclosure, I've encountered cuIBM before; one of my students forked it and developed his own methods using that. However, I haven't contributed to cuIBM itself.)

Where possible, I like to find two reviewers: one close to the domain, and another perhaps coming from a bit further away.

@boyceg @namupatel and/or @arghdos, would you be willing to review this submission to JOSS? @mesnardo, any other suggestions you have would also be appreciated.
