### 2017-06-20 4f7e803
Consider CCFLAGS when linking
