### 2017-06-20 645e6b4
Downgrade yaml-cpp to 0.5.1
With version 0.5.3, the parse would throw an error when not finding a key in the
configuration file (happening when using `const YAML::Node`).
This behavior was not observed with version 0.5.1.

Example:

```c++

int main(int argc, char **argv)
{
  YAML::Node node = YAML::LoadFile("config.yaml");
  std::cout << "a = " << node["a"].as<float>(1) << std::endl;
  std::cout << "b1 = " << node["b"][0].as<float>(2) << std::endl;
  std::cout << "b2 = " << node["b"][1].as<float>(3) << std::endl;

  const YAML::Node node_const = YAML::LoadFile("config.yaml");
  std::cout << "
[…]
