### 2017-06-21 #292 @mesnardo
Hi @kyleniemeyer , thank you for being the editor for this paper.

I just release version 0.1.1 of cuIBM that includes a small bug-fix: downgrading the external library yaml-cpp from version 0.5.3 to 0.5.1.
(yaml-cpp is bundled into cuIBM, in the `external` folder.)
With version 0.5.3, yaml-cpp throws an error when a key parameter is not present in the YAML configuration file, whereas the expected behavior is to use the default value for the parameter.

So, the reviewers should be using the `master` branch of cuIBM or the release 0.1.1.
