### 2017-06-21 #292 @arghdos
@kyleniemeyer I'm willing to review this
