### 2017-06-21 #292 @kyleniemeyer
Awesome, thanks @arghdos! Dan Magee (@osumageed) has also agreed to serve as a second reviewer.
