### 2017-06-21 #301 @kyleniemeyer
@arghdos and @osumageed, this is where the review will take place.

@arghdos since you are the "primary" reviewer (and JOSS is only really set up to handle one reviewer), I'll ask you to check off the boxes at the top. However @osumageed you should also go through and make sure you agree that they are satisfied. Thanks!
