### 2017-07-03 barbagroup/cuIBM#32 @arghdos
JOSS Reviews: Add link to API to Readme
**Functionality documentation**: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

To satisfy the documentation requirement, please add a link to the [Doxygen API docs](http://barbagroup.github.io/cuIBM/) to the readme
