### 2017-07-03 barbagroup/cuIBM#33 @arghdos
JOSS Reviews: Add community guidelines
**Community guidelines**: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

To satisfy the above requirement, please add a [Contribution Guidelines](https://github.com/blog/1184-contributing-guidelines) to cuIBM
