### 2017-07-03 #301 @arghdos
@kyleniemeyer @mesnardo

I have completed a preliminary review of cuIBM.  This is a valuable software that enables CUDA based solution of the 2D incompressible Navier-Stokes equations using immersed boundary methods.  Installation is relatively easy, and is laid out comprehensively in the Readme, there are many examples for use, and there is API level documentation of the doe.

Most of the requirements have been met, but a few issues remain.  Two simple issues have been opened on the cuIBM repo:

1.  [#32](https://github.com/barbagroup/cuIBM/issues/32) - Add a link to the API to the readme for easier access to users.
2.  [#33](https://github.com/barbagroup/cuIBM/issues/33) - Add contribution guidlines, as required for JOSS acceptance (there are examples given in the link left in the issue)

However, there is an open question of on validation of performance.  In one of the early papers linked to in this repo, there are claims of up to 4-8x solution speedup over a single CPU core.  However, it appears the CPU implementation of cuIBM is [not currently functional](https://github.com/barbagroup/cuIBM/issues/20) hence there is no way to validate the performance of cuIBM against any "reference technique".  I am not sure how to proceed with this issue
