### 2017-07-03 barbagroup/cuIBM#34 @OSUmageed
JOSS Reviews: Plots do not work for flying snake example
**Functionality:** The plotting scripts don't work in flying snake examples.

plotForceCoefficients throws an IndexError (same error in 2.7 and 3.6):

```
[info] registering simulation ...
        - description: None
        - software: cuibm
        - directory: /home/osumageed/software/cuIBM/examples/flyingSnake/Re1000AoA30

[info] reading forces ...
/home/osumageed/anaconda/envs/py36/lib/python3.6/site-packages/numpy/core/_methods.py:59: RuntimeWarning: Mean of empty slice.
  warnings.warn("Mean of empty slice.", RuntimeWarning)
/home/osumageed/anaconda/envs/py36/lib/python3.6/site-packages/numpy/core/_methods.py:70: RuntimeWarning: invalid value encountered in double_scalars
  ret = ret.dtype.type(ret / rcount)
Traceback (most recent call last):
  File "scripts/plotForceCoefficients.py", line 15, in <module>
    simulation.get_mean_forces(limits=time_limits)
  File "/home/osumageed/anaconda/envs/py36/lib/python3.6/site-packages/snake-0.3-py3.6.egg/snake/simulation.py", line 106, in get_mean_forces
    order=order)
  File "/home/osumageed/anaconda/envs/py36/lib/python3.6/site-packages/snake-0.3-py3.6.egg/snake/force.py", line 90, in get_mean
    'start': self.times[mask[0]],
IndexError: index 0 is out of bounds for axis 0 with size 0
```

plotVorticity just doesn't make a plot or create images folder.
