### 2017-07-04 #301 @OSUmageed
@kyleniemeyer @arghdos @mesnardo

I completely agree with @arghdos' preliminary review.  And I have a few things to add.

First, I want to say the documentation is really spectacular.  I'm still relatively new to this field, but this is the most well documented source code I've ever read.

I definitely think there needs to be a link to the API in the readme or instructions to build and view it, if only because it allows the user to copy and paste multiple lines of code into a terminal at once which the > symbol in the readme prevents.

The lack of performance validation is an issue.  It doesn't need to be a full example; a study with plots and an adequate description of the hardware that produced them would suffice.  The performance is essential to the **statement of need**.  If it's not faster on the GPU, why not write it in C/C++?

I also opened an issue in the repo [#34](https://github.com/barbagroup/cuIBM/issues/34).  The plotting scripts for the flying snake examples don't work.
