### 2017-07-05 dbccfe2
Add links to documentation in README
Fix issue #32

- Add link to user's documentation.
- Add link to Doxygen API documentation.
