### 2017-07-05 2217e43
Add community guidelines
Should fix issue #33.

Add guidelines on how to contribute to cuIBM.
This includes opening issues and pull-requests, adding new features, submitting
bug fixes, and writting documentation.
