### 2017-07-05 barbagroup/cuIBM#34 @mesnardo
Hi @OSUmageed

Thank you for reporting this problem.

Can you tell for how long you ran the simulation of the snake?
(What is the last time value in the file `forces` located in the simulation directory?)

The python script `plotForceCoefficients.py` is trying to compute the mean forces between 32 and 64 time-units of flow simulation and it looks that it does not find force values within this time range.
If you want, you can either change the time limits or comments some lines directly in the script.

Also, note that the scripts should be run from the root directory of the simulation.
For example:
```bash
cd $CUIBM_DIR/examples/flyingSnakes/Re1000AoA30
python scripts/plotForceCoefficients.py
```

About `plotVorticity.py`, can you provide the error message?
