### 2017-07-05 #301 @mesnardo
Thank you @arghdos for the preliminary review and for your suggestions.

I have added a link to the Doxygen API documentation in the README of cuIBM.
(Note that there is also a link to API documentation at the top of the GitHub page of cuIBM, next to the title.)
I have also added community guidelines in the file `CONTRIBUTING.md`.

Those changes are now available in the branch `joss-reviews-hotfix-0.1.2`.
This branch will be merged into `master` and a new release (v0.1.2) will be created once we have satisfied all your requests.
