### 2017-07-05 2a8ef22
Increase number of time steps
to match the number of time steps used in Krishnan et al. (2014).
The flying-snake examples contains the parameters that were used
in Krishnan et al. (2014).

200,000 time steps with time-step size of 0.0004 to compute the numerical
solution over 80 non-dimensional time units.
