### 2017-07-05 4766f35
Add runtime of simulations
Referring to issue #34.

For long runs, we add a README in the corresponding example directories that
provides the runtime obtained on a specific GPU device with information on how
to change the number of time steps to run or the frequency of saving of the flow
solution.

Fo the flying-snake examples, we also provide information about the time range
used in the Python post-processing script `plotForceCoefficients.py` to
compute the mean force coefficients and Strouhal number.

In the README of the project directory, we also provide the runtimes of most
examples (with
[…]
