### 2017-07-05 barbagroup/cuIBM#34 @mesnardo
You are right, the simulations for the flying snake take a while to complete.

Actually, I pushed a `simParams.yaml` file for only 10 time steps but it should be 200,000 time steps to compute 80 time units of flow simulation (as it was done in Krishnan et al., 2014).
Therefore, I increase the number of time steps to 200,000.

As you suggested, I also placed a README in each snake examples as well as in the flapping-wing example to let the users know that those are long runs. (I am providing the runtime for the GPU device used.)
In those READMEs, I also mention how to change the number of time steps and the frequency of saving.

In the README of the project directory, we now provide the runtime (and the GPU used) for most examples with, again, information on how to change the number of time steps and the saving frequency.
