### 2017-07-06 #301 @kyleniemeyer
Thanks for your reviews @arghdos and @OSUmageed! Do the documentation and community guideline changes that @mesnardo made satisfy those concerns? If so, you can close those issues / check the boxes above.

@mesnardo It seems like the performance claim is the only remaining issue—do you have a response on that?
