### 2017-07-06 #301 @arghdos
@kyleniemeyer I've closed my issues and updated the review
