### 2017-07-06 #301 @OSUmageed
@kyleniemeyer @arghdos @mesnardo I closed my issue as well.
