### 2017-07-06 #301 @mesnardo
@arghdos @OSUmageed Thank you for your reviews and closing the issues.

@kyleniemeyer Concerning the performance and the fact that the CPU routines are not functional, I am waiting for the input from @anushkrish who has run the tests to compare the GPU version of cuIBM to the CPU one.
