### 2017-07-10 #301 @arfon
@kyleniemeyer - is this good to accept now?
