### 2017-07-10 #301 @kyleniemeyer
@arfon no, still waiting on this outstanding performance issue to be resolved
