### 2017-07-18 #301 @labarba
@kyleniemeyer @arghdos There is no outstanding performance issue — the CPU version is not maintained, not supported, and there is a note to that effect in the README. The reviewer is referencing an *old* preprint from many years ago, not *this* paper. We make no CPU performance claims in this paper, nor on the repository of the software. The software repository _only_ includes performance claims in the form of runtime for the examples, on GPU.
