### 2017-07-18 #301 @kyleniemeyer
OK, @labarba @mesnardo, as long as there is no performance claim vs CPU. So, it looks like the software itself is good to go.

However, the [article itself](https://github.com/barbagroup/cuIBM/blob/master/paper.md) has a few outstanding issues:

- [ ] The references in the paper are missing DOIs (where appropriate). Could you add those to the `paper.bib` file?
- [ ] Could you add a more general statement of need perhaps to the beginning of the article, that a broader audience would understand? For example, something about fluid-structure interaction would be appropriate, which isn't currently mentioned.
- [ ] The sentence beginning with "Among them" is missing a verb.

Once those things are resolved, this submission will be ready for publication.
