### 2017-07-18 #301 @labarba
We'll look into that, and we also just discussed with @mesnardo that we'll remove the CPU routines from the master branch and leave it in a `deprecated` branch.

BTW, this software does _not_ compute (two-way) fluid-structure interaction (only prescribed body motions).
