### 2017-07-18 #301 @labarba
(And, as you heard me rant at SciPy, we really have to _stop_ talking about GPU vs. CPU comparisons. This is a fixation that is an unfortunate consequence of the vendor marketing, and it no longer makes sense.)
