### 2017-07-18 c1d7399
Add missing DOIs
