+++ paper.bib
@@ -6,7 +6,8 @@
   number={2},
   pages={2118--2137},
   year={2007},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.jcp.2007.03.005}
 }
 
 @article{FadlunEtAl2000,
@@ -17,7 +18,8 @@
   number={1},
   pages={35--60},
   year={2000},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1006/jcph.2000.6484}
 }
 
 @article{Perot1993,
@@ -28,7 +30,8 @@
   number={1},
   pages={51--58},
   year={1993},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1006/jcph.1993.1162}
 }
 
 @article{KrishnanEtAl2014,
@@ -39,7 +42,8 @@
   number={3},
   pages={031901},
   year={2014},
-  publisher={AIP}
+  publisher={AIP},
+  doi={10.1063/1.4866444}
 }
 
 @phdthesis{Krishnan2015,
