### 2017-07-18 d900ffe
Review paper for JOSS
Add general explanation about immersed-boundary method.
Fix typos (remove hyphen, spell "study" correctly).
Add missing verb.
