### 2017-07-18 4fbf8e6
Remove CPU routines
The code with the non-functional CPU routines is now available in a separate
branch named `deprecated-cpu-routines`.

Remove section "Known issues" from README (and open issue for that).
