+++ src/solvers/kernels/generateE.h
@@ -17,11 +17,6 @@
 namespace kernels
 {
 
-// generate the interpolation matrix (on the host)
-void generateEHost(int *ERows,  int *ECols,  real *EVals,
-                   int nx, int ny, real *x, real *y, real *dx,
-                   int totalPoints, real *xB, real *yB, int *I, int *J);
-
 // compute elements of the interpolation matrix
 __global__
 void generateE(int *ERows,  int *ECols,  real *EVals,
