### 2017-07-18 #301 @mesnardo
@kyleniemeyer , we have removed the CPU routines, there are now in a separate branch called `deprecated-cpu-routines`.
Also, we have updated the paper and the references with the suggestions you made above.
