### 2017-07-18 #301 @kyleniemeyer
OK, looks good! One last thing: the Known Issues at the bottom still says "CPU routines do not work." Could you correct that, and then provide a DOI for the archived software here?
