### 2017-07-18 #301 @mesnardo
You should be looking at the branch `joss-reviews-hotfix-0.1.2`; it includes all the changes made during the peer-review.
We are waiting for acceptance before making the new archive and getting a DOI.
