### 2017-07-18 #301 @kyleniemeyer
@mesnardo ah, got it—I was indeed looking at the master README.

OK, this is accepted for publication, so please merge that branch and archive the software.
