### 2017-07-19 85c00bb
[v0.1.2] Update Doxygen API documentation
