+++ search/all_6.js
@@ -2,42 +2,46 @@ var searchData=
 [
   ['g',['G',['../classFadlunEtAlSolver.html#a03d378fadc6775d1a8a0eb8aafdd5559',1,'FadlunEtAlSolver']]],
   ['gamma',['gamma',['../classintegrationScheme.html#a8acd5cd4712a28053def08e46ad87719',1,'integrationScheme']]],
-  ['generatea',['generateA',['../classDirectForcingSolver.html#adf433f1ffa4f2f83b4f5240adbcdb2ac',1,'DirectForcingSolver::generateA()'],['../classNavierStokesSolver.html#aa695d86ee831519ec5f97a60ac9851f3',1,'NavierStokesSolver::generateA()'],['../classDirectForcingSolver.html#a2fd346de710f7d1e567d5687ac29efe1',1,'DirectForcingSolver::generateA(real alpha)'],['../classDirectForcingSolver.html#a111042e3a1a8656efae16d5f2aa53c99',1,'DirectForcingSolver::generateA(real alpha)'],['../classNavierStokesSolver.html#abce7e393ec94c45efb40110660307e0a',1,'NavierStokesSolver::generateA(real alpha)'],['../classNavierStokesSolver.html#ac54854c8fd23e326f12d07611b089074',1,'NavierStokesSolver::generateA(real alpha)'],['../namespacekernels.html#a419ce463ea9f66315cdaf9cf82150018',1,'kernels::generateA()']]],
+  ['generatea',['generateA',['../classDirectForcingSolver.html#adf433f1ffa4f2f83b4f5240adbcdb2ac',1,'DirectForcingSolver::generateA()'],['../classNavierStokesSolver.html#aa695d86ee831519ec5f97a60ac9851f3',1,'NavierStokesSolver::generateA()'],['../classDirectForcingSolver.html#a2fd346de710f7d1e567d5687ac29efe1',1,'DirectForcingSolver::generateA()'],['../classNavierStokesSolver.html#abce7e393ec94c45efb40110660307e0a',1,'NavierStokesSolver::generateA()'],['../namespacekernels.html#a419ce463ea9f66315cdaf9cf82150018',1,'kernels::generateA()']]],
   ['generatea_2ecu',['generateA.cu',['../generateA_8cu.html',1,'']]],
   ['generatea_2eh',['generateA.h',['../generateA_8h.html',1,'']]],
-  ['generatea_2einl',['generateA.inl',['../DirectForcing_2generateA_8inl.html',1,'(Global Namespace)'],['../NavierStokes_2generateA_8inl.html',1,'(Global Namespace)']]],
+  ['generatea_2einl',['generateA.inl',['../DirectForcing_2generateA_8inl.html',1,'']]],
+  ['generatea_2einl',['generateA.inl',['../NavierStokes_2generateA_8inl.html',1,'']]],
   ['generateadirectforcing',['generateADirectForcing',['../namespacekernels.html#a8b844498efcbc30574de8dbcdbe3242f',1,'kernels']]],
-  ['generatebc1',['generateBC1',['../classNavierStokesSolver.html#ac1f4a49ace3e55641e645a00a07c4992',1,'NavierStokesSolver::generateBC1()'],['../classNavierStokesSolver.html#a1999ecf358bfcc19adaedd4c74640193',1,'NavierStokesSolver::generateBC1()'],['../classNavierStokesSolver.html#aff46cf92faeffb667960f188321afde5',1,'NavierStokesSolver::generateBC1()']]],
[…]
