+++ search/functions_2.js
@@ -3,8 +3,8 @@ var searchData=
   ['calculateboundingboxes',['calculateBoundingBoxes',['../classbodies.html#aa6784a8a4cd35d5ebac250e7e1b09043',1,'bodies']]],
   ['calculatecellindices',['calculateCellIndices',['../classbodies.html#a96456bdc053210c785bde1598b78bd90',1,'bodies']]],
   ['calculateexplicitlambdaterms',['calculateExplicitLambdaTerms',['../classDFModifiedSolver.html#a69604835337d81c7d23884d23b2ba30f',1,'DFModifiedSolver::calculateExplicitLambdaTerms()'],['../classFadlunEtAlSolver.html#a1e727a1fdd3cfe045644ebcc04d96f1f',1,'FadlunEtAlSolver::calculateExplicitLambdaTerms()'],['../classNavierStokesSolver.html#a116981a2f6decd1b870959619ee83e4a',1,'NavierStokesSolver::calculateExplicitLambdaTerms()']]],
-  ['calculateexplicitqterms',['calculateExplicitQTerms',['../classNavierStokesSolver.html#ace2b1b03f52345fea182dbcf4e7550b1',1,'NavierStokesSolver::calculateExplicitQTerms()'],['../classNavierStokesSolver.html#a87a31652151ecd04392b0e0f0d9f8c77',1,'NavierStokesSolver::calculateExplicitQTerms()'],['../classNavierStokesSolver.html#af847c45742e2083eef2ea406a5881459',1,'NavierStokesSolver::calculateExplicitQTerms()']]],
-  ['calculateforce',['calculateForce',['../classNSWithBody.html#a478f1697f038e849522adea85784136b',1,'NSWithBody::calculateForce()'],['../classTairaColoniusSolver.html#a1766f3af7c6cefacfc9d990ba69baaee',1,'TairaColoniusSolver::calculateForce()'],['../classNSWithBody.html#a86cfcd44e000157d69959df20b67220d',1,'NSWithBody::calculateForce()'],['../classNSWithBody.html#a4397d6acdf54693b3a58fcdeff4d6973',1,'NSWithBody::calculateForce()']]],
+  ['calculateexplicitqterms',['calculateExplicitQTerms',['../classNavierStokesSolver.html#ace2b1b03f52345fea182dbcf4e7550b1',1,'NavierStokesSolver::calculateExplicitQTerms()'],['../classNavierStokesSolver.html#a87a31652151ecd04392b0e0f0d9f8c77',1,'NavierStokesSolver::calculateExplicitQTerms()']]],
+  ['calculateforce',['calculateForce',['../classNSWithBody.html#a478f1697f038e849522adea85784136b',1,'NSWithBody::calculateForce()'],['../classTairaColoniusSolver.html#a1766f3af7c6cefacfc9d990ba69baaee',1,'TairaColoniusSolver::calculateForce()'],['../classNSWithBody.html#a4397d6acdf54693b3a58fcdeff4d6973',1,'NSWithBody::calculateForce()']]],
   ['commandlineparse1',['commandLineParse1',['../namespaceio.html#ac0c5d8b6cb36f9d16d0e91bf45d0eaec',1,'io']]],
   ['commandlineparse2',['commandLineParse2',['../namespaceio.html#abb630efc2279debb4199445ddefba65b',1,'io']]],
   ['convectiontermu',['convectionTermU',['../namespacekernels.html#acef6095fe398e00a034bccfd1cb9d395',1,'kernels']]],
