### 2017-07-19 #301 @mesnardo
@kyleniemeyer Branch `joss-reviews-hotfix-0.1.2` has been merged into `master`, cuIBM-0.1.2 has been released and uploaded to [Zenodo](https://doi.org/10.5281/zenodo.832338) (10.5281/zenodo.832338).
