### 2017-07-21 #301 @mesnardo
Yay!
Thank you very much @kyleniemeyer , @arghdos , and @OSUmageed for your review of cuIBM and your useful suggestions!
